refactor: 온보딩 단계 컴포넌트 구조 변경 및 새로운 컴포넌트 추가
