+++ omechu-app/src/app/onboarding/components/ConstitutionStep.tsx
@@ -2,9 +2,9 @@
 
 import React from "react";
 
+import ListButton from "@/app/components/common/button/ListButton";
 import { useOnboardingStore } from "@/lib/stores/onboarding.store";
 
-import OnboardingButton from "./OnboardingButton";
 import OnboardingStepLayout from "./OnboardingStepLayout";
 
 const ConstitutionStep = () => {
@@ -13,19 +13,20 @@ const ConstitutionStep = () => {
     "감기에 잘 걸리는 편이에요",
     "소화가 잘 안 되는 날이 많아요",
     "열이 많아서 더위를 잘 타요",
-    "추위를 잘 타고 몸이 쉽게 차가워져요",
+    "추위를 잘 타고 몸이 쉽 게 차가워져요",
   ];
 
   return (
     <OnboardingStepLayout title="체질은 무엇인가요?" paddingClassName="px-0">
       {options.map((option) => (
-        <OnboardingButton
+        <ListButton
           key={option}
           onClick={() => toggleConstitution(option)}
           isSelected={constitution.includes(option)}
+          textSize="base"
         >
           {option}
-        </OnboardingButton>
+        </ListButton>
       ))}
     </OnboardingStepLayout>
   );
