+++ omechu-app/src/app/onboarding/components/GenderStep.tsx
@@ -2,9 +2,9 @@
 
 import React from "react";
 
+import ListButton from "@/app/components/common/button/ListButton";
 import { useOnboardingStore } from "@/lib/stores/onboarding.store";
 
-import OnboardingButton from "./OnboardingButton";
 import OnboardingStepLayout from "./OnboardingStepLayout";
 
 const GenderStep = () => {
@@ -17,18 +17,18 @@ const GenderStep = () => {
   return (
     <OnboardingStepLayout title="성별은 무엇인가요?">
       <div className="flex w-full gap-4">
-        <OnboardingButton
+        <ListButton
           onClick={() => handleGenderSelect("여성")}
           isSelected={gender === "여성"}
         >
           여성
-        </OnboardingButton>
-        <OnboardingButton
+        </ListButton>
+        <ListButton
           onClick={() => handleGenderSelect("남성")}
           isSelected={gender === "남성"}
         >
           남성
-        </OnboardingButton>
+        </ListButton>
       </div>
     </OnboardingStepLayout>
   );
