+++ omechu-app/src/app/onboarding/components/OnboardingStepLayout.tsx
