+++ omechu-app/src/app/onboarding/components/PreferredFoodStep.tsx
@@ -2,9 +2,9 @@
 
 import React from "react";
 
+import ListButton from "@/app/components/common/button/ListButton";
 import { useOnboardingStore } from "@/lib/stores/onboarding.store";
 
-import OnboardingButton from "./OnboardingButton";
 import OnboardingStepLayout from "./OnboardingStepLayout";
 
 const PreferredFoodStep = () => {
@@ -14,13 +14,13 @@ const PreferredFoodStep = () => {
   return (
     <OnboardingStepLayout title={"평소 자주 먹거나 좋아하는 \n 음식이 있나요?"}>
       {options.map((option) => (
-        <OnboardingButton
+        <ListButton
           key={option}
           onClick={() => togglePreferredFood(option)}
           isSelected={preferredFood.includes(option)}
         >
           {option}
-        </OnboardingButton>
+        </ListButton>
       ))}
     </OnboardingStepLayout>
   );
